# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atzicay API",
        "description": "API documentation for Atzicay",
        "contact": {
            "email": "atzicay@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api/atzicay/v1/"
        }
    ],
    "paths": {
...
        "/game-instances": {
            "get": {
                "tags": [
                    "GameInstances"
                ],
                "summary": "Get all game instances",
                "description": "Retrieves all game instances.",
                "operationId": "6a6eb13a560ea44e226546518c7e400f",
                "responses": {
                    "200": {
                        "description": "List of game instances",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GameInstances"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No game instances found"
                    }
                }
            },
            "post": {
                "tags": [
                    "GameInstances"
                ],
                "summary": "Create a new game instance",
                "description": "Creates a new game instance.",
                "operationId": "f801184436c315aa5c7a3ed798ae8e07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameInstancesDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game instance created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInstances"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game-instances/{id}": {
            "get": {
                "tags": [
                    "GameInstances"
                ],
                "summary": "Get game instance by ID",
                "description": "Retrieves a game instance by its ID.",
                "operationId": "0acfee30593d526c028ba5d24be8be5b",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the game instance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game instance details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInstances"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game instance not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "GameInstances"
                ],
                "summary": "Update a game instance",
                "description": "Updates an existing game instance.",
                "operationId": "9b6c205e8523cb877774dba0099a8175",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the game instance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameInstancesDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game instance updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInstances"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "GameInstances"
                ],
                "summary": "Delete a game instance",
                "description": "Deletes a game instance by its ID.",
                "operationId": "d54aed7c9f517aed29f79a70ce9190fa",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the game instance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game instance deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInstances"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game-progress/{id}": {
            "get": {
                "tags": [
                    "GameProgress"
                ],
                "summary": "Get Game Progress by ID",
                "description": "Retrieves a Game Progress by its ID.",
                "operationId": "ee7f904bc1970ac0bd21fc8754c14820",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the Game Progress to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game Progress found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameProgress"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game Progress not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "GameProgress"
                ],
                "summary": "Update Game Progress",
                "description": "Updates an existing Game Progress.",
                "operationId": "278ed5fb6ffc5857d882e62d089c8192",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "ID of the Game Progress to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameProgressDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game Progress updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameProgress"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game-progress": {
            "post": {
                "tags": [
                    "GameProgress"
                ],
                "summary": "Create Game Progress",
                "description": "Creates a new Game Progress.",
                "operationId": "51e3af1567ad7b937333682e1db571d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameProgressDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game Progress created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameProgress"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GameInstancesDTO": {
                "title": "Game Instances DTO",
                "description": "Data transfer object for game instances",
                "properties": {
                    "Name": {
                        "description": "The name of the game instance",
                        "type": "string"
                    },
                    "Description": {
                        "description": "The description of the game instance",
                        "type": "string"
                    },
                    "ProfessorId": {
                        "description": "The ID of the professor associated with the game instance",
                        "type": "integer"
                    },
                    "Activated": {
                        "description": "Indicates if the game instance is active",
                        "type": "boolean"
                    },
                    "Difficulty": {
                        "description": "The difficulty level of the game instance",
                        "type": "string",
                        "enum": [
                            "E",
                            "M",
                            "H"
                        ]
                    },
                    "Visibility": {
                        "description": "Indicates if the game instance is visible",
                        "type": "string",
                        "enum": [
                            "P",
                            "R"
                        ]
                    }
                },
                "type": "object"
            },
            "GameProgressDTO": {
                "title": "Game Progress DTO",
                "description": "Data transfer object for game progress",
                "properties": {
                    "GameSessionId": {
                        "description": "The ID of the game session",
                        "type": "integer"
                    },
                    "Progress": {
                        "description": "The progress of the game session",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "GameInstances": {
                "required": [
                    "id",
                    "name",
                    "description",
                    "professorId",
                    "activated",
                    "difficulty",
                    "visibility"
                ],
                "properties": {
                    "Id": {
                        "type": "integer",
                        "example": 1
                    },
                    "Name": {
                        "type": "string",
                        "example": "Game Instance 1"
                    },
                    "Description": {
                        "type": "string",
                        "example": "This is a description of the game instance."
                    },
                    "ProfessorId": {
                        "type": "integer",
                        "example": 123
                    },
                    "Activated": {
                        "type": "boolean",
                        "example": true
                    },
                    "Difficulty": {
                        "type": "string",
                        "enum": [
                            "E",
                            "M",
                            "H"
                        ],
                        "example": "medium"
                    },
                    "Visibility": {
                        "type": "string",
                        "enum": [
                            "P",
                            "R"
                        ],
                        "example": "public"
                    }
                },
                "type": "object"
            },
            "GameProgress": {
                "required": [
                    "id",
                    "gameSessionId",
                    "progress"
                ],
                "properties": {
                    "Id": {
                        "type": "integer",
                        "example": 1
                    },
                    "GameSessionId": {
                        "type": "integer",
                        "example": 123
                    },
                    "Progress": {
                        "type": "string",
                        "example": "Level 1 completed"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}